+++ veridui/src/main/java/com/appliedrec/verid/ui/IVideoRecorder.java
@@ -1,0 +1,17 @@
+package com.appliedrec.verid.ui;
+
+import java.io.File;
+
+/**
+ * Video recorder interface
+ * @since 1.17.0
+ */
+public interface IVideoRecorder {
+
+    /**
+     * Get the file that holds the video recording
+     * @return File
+     * @since 1.17.0
+     */
+    File getVideoFile();
+}